# static/swagger.json
{
  "info": {
    "title": "SmartOven",
    "description": "Oven. But smart.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
    "/recipe/find": {
      "post": {
        "summary": "Search a new recipe and add it in the oven's database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Banana bread"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully found and added a new recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Test recipe"
                    },
                    "prep_time": {
                      "type": "integer",
                      "example": 10
                    },
                    "prep_details": {
                      "type": "string",
                      "example": "Test recipe details"
                    },
                    "baking_temperature": {
                      "type": "integer",
                      "example": 20
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required fields for a recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing fields to search recipe"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorised client",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The client is not authorised to search"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No recipe with name found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Duplicate recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "A recipe with the same name already exists"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/recipe": {
      "get": {
        "summary": "Get all recipes that are saved in the oven's database",
        "responses": {
          "200": {
            "description": "Successfully returned a list of recipes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "name"
                    ],
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "prep_time": {
                        "type": "integer"
                      },
                      "prep_details": {
                        "type": "string"
                      },
                      "baking_temperature": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new recipe in the oven's database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "prep_time",
                  "prep_details",
                  "baking_temperature"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Test recipe"
                  },
                  "prep_time": {
                    "type": "integer",
                    "example": 10
                  },
                  "prep_details": {
                    "type": "string",
                    "example": "Test recipe details"
                  },
                  "baking_temperature": {
                    "type": "integer",
                    "example": 20
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully added a new recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully added the recipe"
                    },
                    "id": {
                      "type": "string",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required fields for a recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing fields to perform adding of a recip"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Duplicate recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "A recipe with the same name already exists"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/recipe/{recipe_name}": {
      "get": {
        "summary": "Get recipe with the specified name from oven's database",
        "responses": {
          "200": {
            "description": "Successfully returned the specified recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name"
                  ],
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "prep_time": {
                      "type": "integer"
                    },
                    "prep_details": {
                      "type": "string"
                    },
                    "baking_temperature": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing recipe name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Recipe not found"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "recipe_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/recipe/{recipe_id}": {
      "put": {
        "summary": "Update a recipe with the specified id from oven's database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "prep_time": {
                    "type": "integer"
                  },
                  "prep_details": {
                    "type": "string"
                  },
                  "baking_temperature": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully returned the specified recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully updated recipe with specified id"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Recipe is missing or has unallowed fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Request content contains unallowed fields"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "A recipe with the specified id does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error during recipe update",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occured during update of the recipe"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "summary": "Delete a recipe with the specified id from oven's database",
        "responses": {
          "200": {
            "description": "Successfully returned the specified recipe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted recipe with specified id"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recipe not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "A recipe with the specified id does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "definitions": {},
  "openapi": "3.0.2"
}